feat: main 페이지 desktop UI 구현
