+++ src/app/(routes)/(tabs)/main/page.tsx
@@ -1,194 +1,20 @@
-import { Button } from '@/components/ui/button'
+import PageTitle from '@/components/page-title'
 import icons from '@/constants/icons'
-import Image from 'next/image'
+import QuizBanner from './components/quiz-banner'
+import QuizMaker from './components/quiz-maker'
+import Achievements from './components/achievements'
 
 export default function Main() {
   return (
-    <main className="flex flex-col px-[20px] pb-[40px] lg:gap-[28px]">
-      <div className="mb-[26px] flex items-start gap-[8px]">
-        <h1 className="text-h3-bold text-gray-09">파워업 퀴즈</h1>
-        <Image src={icons.powerUpQuiz} width={26} height={26} alt="" />
-      </div>
+    <main className="flex flex-col px-[20px] pb-[40px]">
+      <PageTitle title="파워업 퀴즈" icon={icons.powerUpQuiz} />
 
-      <section className="mb-[56px] flex flex-col gap-[24px]">
-        <div className="relative flex min-h-[240px] w-full flex-col justify-between rounded-[12px] bg-orange-02 p-[20px]">
-          <div className="w-[calc(100%-160px)]">
-            <div className="mb-[12px] text-body1-bold-eng text-orange-06">TODAY&apos;s QUIZ</div>
-            <div className="mb-[39px] flex flex-col gap-[8px]">
-              <div className="text-h4-bold text-gray-09">픽토스님을 위한 퀴즈가 준비되었어요</div>
-              <div className="text-body2-medium text-gray-08">4월 25일 목요일</div>
-            </div>
-          </div>
-
-          <Image src={icons.quizReady} width={148} className="absolute right-[18px]" alt="" />
-
-          <Button className="flex w-full gap-[8px] rounded-[8px]">
-            <div>오늘의 퀴즈 시작하기</div>
-            <Image src={icons.arrowRight} width={20.25} height={13.5} alt="" />
-          </Button>
-        </div>
-
-        <div className="flex h-[80px] gap-[7px]">
-          <div className="flex min-w-[164px] flex-1 justify-center gap-[26px] rounded-[12px] bg-white">
-            <div className="flex flex-col gap-[4px] pt-[16px]">
-              <div className="text-body2-medium text-gray-08">퀴즈 연속일</div>
-              <div className="text-h4-bold text-gray-08">
-                <span className="text-orange-06">25</span>일
-              </div>
-            </div>
-            <Image src={icons.calendar} width={40} height={40} alt="" />
-          </div>
-          <div className="flex min-w-[164px] flex-1 justify-center gap-[26px] rounded-[12px] bg-white">
-            <div className="flex flex-col gap-[4px] pt-[16px]">
-              <div className="text-body2-medium text-gray-08">나의 별</div>
-              <div className="text-h4-bold text-gray-08">
-                <span className="text-orange-06">16</span>개
-              </div>
-            </div>
-            <Image src={icons.star} width={40} height={40} alt="" />
-          </div>
-        </div>
+      <section className="mb-[56px] flex flex-col gap-[24px] lg:flex-row lg:gap-[10px]">
+        <QuizBanner />
+        <Achievements />
       </section>
 
-      <section className="flex flex-col gap-[24px]">
-        <div>
-          <div className="flex gap-[8px]">
-            <div className="text-h4-bold text-gray-08">퀴즈 만들기</div>
-            <Image src={icons.circleQuestion} width={20} height={20} alt="" />
-          </div>
-          <div className="text-body2-regular text-gray-07">
-            원하는 노트에서 퀴즈를 만들고 풀어보세요
-          </div>
-        </div>
-
-        <div className="flex flex-col gap-[16px]">
-          <div className="w-full rounded-[12px]">
-            <div className="flex h-[120px] items-center justify-center bg-blue-01">
-              <Icon1 />
-            </div>
-            <div className="bg-white px-[24px] pb-[18px] pt-[16px]">
-              <div className="flex items-center gap-[8px] text-h4-bold text-gray-09">객관식</div>
-              <div className="text-body2-medium text-gray-06">
-                <span>4가지 선택지 중 정답을 고르는 퀴즈</span>
-              </div>
-            </div>
-          </div>
-          <div className="w-full rounded-[12px]">
-            <div className="flex h-[120px] items-center justify-center bg-blue-01">
-              <Icon2 />
-            </div>
-            <div className="bg-white px-[24px] pb-[18px] pt-[16px]">
-              <div className="flex items-center gap-[8px] text-h4-bold text-gray-09">O/X</div>
-              <div className="text-body2-medium text-gray-06">
-                <span>참인지 거짓인지 판단하는 양자택일 퀴즈</span>
-              </div>
-            </div>
-          </div>
-          <div className="w-full rounded-[12px]">
-            <div className="flex h-[120px] items-center justify-center bg-blue-01">
-              <Icon3 />
-            </div>
-            <div className="bg-white px-[24px] pb-[18px] pt-[16px]">
-              <div className="flex items-center gap-[8px] text-h4-bold text-gray-09">
-                <span>빈칸 채우기</span>
-                <span className="block h-fit rounded-[3px] bg-gray-02 px-[6px] pb-[2px] pt-px text-[10px] text-gray-08">
-                  Coming soon
-                </span>
-              </div>
-              <div className="text-body2-medium text-gray-06">
-                주어진 문장의 빈 곳을 채우는 퀴즈
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
+      <QuizMaker />
     </main>
-  )
-}
-
-/** TODO: Icon 분리 및 모션 적용 */
-function Icon1() {
-  return (
-    <svg
-      width="127"
-      height="80"
-      viewBox="0 0 127 80"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <rect x="21.3516" y="0.800049" width="77.8174" height="17.0652" rx="3.89087" fill="#D7E2FF" />
-      <rect x="21.3516" y="20.9031" width="77.8174" height="17.0652" rx="3.89087" fill="#D7E2FF" />
-      <rect x="29.7812" y="41.0056" width="77.8174" height="17.0652" rx="3.89087" fill="#95B0F8" />
-      <rect x="21.3516" y="61.1086" width="77.8174" height="17.0652" rx="3.89087" fill="#D7E2FF" />
-      <circle cx="39.4155" cy="49.5481" r="5.74366" fill="#F5F7F9" />
-      <path
-        d="M37.1182 49.4738L38.8413 51.1969L42.2875 47.7507"
-        stroke="#577CFF"
-        strokeWidth="1.14873"
-        strokeLinecap="round"
-      />
-      <path
-        d="M108.759 67.1532L101.974 56.6385C101.04 55.1913 102.201 53.3079 103.914 53.4921L116.255 54.8197C118.04 55.0117 118.732 57.2444 117.367 58.4117L114.291 61.0441C113.975 61.3148 113.749 61.6755 113.642 62.0781L112.46 66.5649C112.006 68.2861 109.724 68.6488 108.759 67.1532Z"
-        fill="#FF9100"
-      />
-    </svg>
-  )
-}
-
-function Icon2() {
-  return (
-    <svg
-      width="176"
-      height="112"
-      viewBox="0 0 176 112"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <rect x="90.0664" y="28.8" width="76.5556" height="54.1778" rx="9.42222" fill="#95B0F8" />
-      <rect x="20" y="32.8" width="64.8" height="45.8585" rx="7.97538" fill="#D7E2FF" />
-      <circle cx="52.8993" cy="55.7291" r="11.9631" stroke="#95B0F8" strokeWidth="3.98769" />
-      <path
-        d="M117.155 44.1113L139.533 66.4891"
-        stroke="#F0F4FF"
-        strokeWidth="5.88889"
-        strokeLinecap="round"
-      />
-      <path
-        d="M139.533 44.1113L117.155 66.4891"
-        stroke="#F0F4FF"
-        strokeWidth="5.88889"
-        strokeLinecap="round"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M124.628 20.9633C125.738 20.8553 126.862 20.8 128 20.8C129.137 20.8 130.262 20.8553 131.371 20.9633C132.089 21.0331 132.615 21.6719 132.545 22.39C132.475 23.1081 131.836 23.6335 131.118 23.5636C130.092 23.4638 129.052 23.4127 128 23.4127C126.947 23.4127 125.907 23.4638 124.881 23.5636C124.163 23.6335 123.524 23.1081 123.454 22.39C123.385 21.6719 123.91 21.0331 124.628 20.9633ZM119.64 23.1444C119.849 23.8349 119.459 24.5642 118.769 24.7734C116.759 25.3823 114.833 26.1849 113.013 27.1597C112.377 27.5004 111.586 27.2609 111.245 26.6249C110.904 25.989 111.144 25.1973 111.78 24.8566C113.75 23.8015 115.835 22.9324 118.011 22.273C118.701 22.0638 119.431 22.454 119.64 23.1444ZM136.359 23.1444C136.568 22.454 137.298 22.0638 137.988 22.273C140.165 22.9324 142.249 23.8015 144.219 24.8566C144.855 25.1972 145.095 25.989 144.754 26.6249C144.414 27.2609 143.622 27.5003 142.986 27.1597C141.166 26.1849 139.24 25.3823 137.231 24.7734C136.54 24.5642 136.15 23.8349 136.359 23.1444ZM147.985 28.788C148.443 28.2306 149.266 28.15 149.824 28.608C151.564 30.038 153.162 31.6355 154.592 33.376C155.05 33.9335 154.969 34.7567 154.412 35.2147C153.854 35.6727 153.031 35.5921 152.573 35.0346C151.251 33.4256 149.774 31.9487 148.165 30.6267C147.608 30.1686 147.527 29.3455 147.985 28.788ZM108.014 28.788C108.472 29.3455 108.392 30.1687 107.834 30.6267C106.225 31.9487 104.748 33.4256 103.426 35.0346C102.968 35.5921 102.145 35.6727 101.588 35.2147C101.03 34.7567 100.95 33.9335 101.408 33.376C102.838 31.6355 104.435 30.038 106.176 28.608C106.733 28.15 107.556 28.2306 108.014 28.788ZM156.575 38.4455C157.211 38.1048 158.002 38.3442 158.343 38.9802C159.398 40.9502 160.267 43.0351 160.927 45.2114C161.136 45.9019 160.746 46.6312 160.055 46.8404C159.365 47.0496 158.635 46.6595 158.426 45.969C157.817 43.9594 157.015 42.0338 156.04 40.2138C155.699 39.5778 155.939 38.7861 156.575 38.4455ZM99.4245 38.4455C100.06 38.7861 100.3 39.5778 99.9593 40.2138C98.9845 42.0338 98.1819 43.9594 97.573 45.969C97.3638 46.6595 96.6345 47.0496 95.944 46.8404C95.2535 46.6312 94.8634 45.9019 95.0726 45.2114C95.7319 43.0351 96.601 40.9502 97.6562 38.9802C97.9968 38.3443 98.7885 38.1048 99.4245 38.4455ZM95.1895 50.6549C95.9076 50.7247 96.4331 51.3635 96.3632 52.0816C96.2634 53.1072 96.2123 54.1475 96.2123 55.2001C96.2123 56.2526 96.2634 57.2929 96.3632 58.3185C96.4331 59.0366 95.9076 59.6754 95.1895 59.7452C94.4715 59.8151 93.8327 59.2897 93.7628 58.5716C93.6548 57.462 93.5996 56.3373 93.5996 55.2001C93.5996 54.0628 93.6548 52.9381 93.7628 51.8285C93.8327 51.1104 94.4715 50.585 95.1895 50.6549ZM160.81 50.6549C161.528 50.585 162.167 51.1104 162.236 51.8285C162.344 52.9381 162.4 54.0628 162.4 55.2001C162.4 56.3373 162.344 57.462 162.236 58.5716C162.167 59.2897 161.528 59.8151 160.81 59.7452C160.092 59.6754 159.566 59.0366 159.636 58.3185C159.736 57.2929 159.787 56.2526 159.787 55.2001C159.787 54.1475 159.736 53.1072 159.636 52.0816C159.566 51.3635 160.092 50.7247 160.81 50.6549ZM95.944 63.5597C96.6345 63.3505 97.3638 63.7406 97.573 64.4311C98.1819 66.4407 98.9845 68.3663 99.9593 70.1863C100.3 70.8223 100.06 71.614 99.4245 71.9546C98.7885 72.2953 97.9968 72.0558 97.6562 71.4199C96.601 69.4499 95.7319 67.365 95.0726 65.1887C94.8634 64.4982 95.2535 63.7689 95.944 63.5597ZM160.055 63.5597C160.746 63.7689 161.136 64.4982 160.927 65.1887C160.267 67.365 159.398 69.4499 158.343 71.4199C158.002 72.0558 157.211 72.2953 156.575 71.9546C155.939 71.614 155.699 70.8223 156.04 70.1863C157.015 68.3663 157.817 66.4407 158.426 64.4311C158.635 63.7406 159.365 63.3505 160.055 63.5597ZM154.412 75.1854C154.969 75.6434 155.05 76.4666 154.592 77.0241C153.162 78.7646 151.564 80.3621 149.824 81.7921C149.266 82.2501 148.443 82.1695 147.985 81.6121C147.527 81.0546 147.608 80.2314 148.165 79.7734C149.774 78.4514 151.251 76.9745 152.573 75.3655C153.031 74.808 153.854 74.7274 154.412 75.1854ZM101.588 75.1854C102.145 74.7274 102.968 74.808 103.426 75.3655C104.748 76.9745 106.225 78.4514 107.834 79.7734C108.392 80.2314 108.472 81.0546 108.014 81.6121C107.556 82.1695 106.733 82.2501 106.176 81.7921C104.435 80.3621 102.838 78.7646 101.408 77.0241C100.95 76.4666 101.03 75.6434 101.588 75.1854ZM111.245 83.7752C111.586 83.1392 112.377 82.8997 113.013 83.2404C114.833 84.2152 116.759 85.0178 118.769 85.6267C119.459 85.8359 119.849 86.5652 119.64 87.2557C119.431 87.9461 118.701 88.3363 118.011 88.1271C115.835 87.4677 113.75 86.5986 111.78 85.5435C111.144 85.2029 110.904 84.4111 111.245 83.7752ZM144.754 83.7752C145.095 84.4111 144.855 85.2028 144.219 85.5435C142.249 86.5986 140.165 87.4677 137.988 88.1271C137.298 88.3363 136.568 87.9461 136.359 87.2557C136.15 86.5652 136.54 85.8359 137.231 85.6267C139.24 85.0178 141.166 84.2152 142.986 83.2404C143.622 82.8997 144.414 83.1392 144.754 83.7752ZM123.454 88.0101C123.524 87.292 124.163 86.7666 124.881 86.8365C125.907 86.9363 126.947 86.9874 128 86.9874C129.052 86.9874 130.092 86.9363 131.118 86.8365C131.836 86.7666 132.475 87.292 132.545 88.0101C132.615 88.7282 132.089 89.367 131.371 89.4368C130.262 89.5448 129.137 89.6001 128 89.6001C126.862 89.6001 125.738 89.5448 124.628 89.4368C123.91 89.3669 123.385 88.7282 123.454 88.0101Z"
-        fill="#FF9100"
-      />
-    </svg>
-  )
-}
-
-function Icon3() {
-  return (
-    <svg
-      width="176"
-      height="112"
-      viewBox="0 0 176 112"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <rect x="15.2002" y="40.8" width="146.044" height="32" rx="4.71111" fill="#D7E2FF" />
-      <path
-        d="M28.8857 53.9536H36.3353V55.208H35.1961V58.9968H36.5017V60.264H28.7321V58.9968H30.0505V55.208H28.8857V53.9536ZM31.6121 58.9968H33.6217V55.208H31.6121V58.9968ZM39.7532 53.9408H46.998V55.208H41.3404V56.3856H46.8956V57.64H41.3404V58.9072H47.1004V60.1744H39.7532V53.9408ZM53.4111 53.8H54.9983V55.5152C54.9983 57.192 56.0479 58.6128 58.3647 59.0992L57.6095 60.3536C55.9711 59.9568 54.8319 59.0992 54.2047 57.9216C53.5775 59.0992 52.4383 59.9568 50.7999 60.3536L50.0447 59.0992C52.3487 58.6128 53.4111 57.192 53.4111 55.5152V53.8ZM61.5522 53.8384H68.7202V60.264H67.1202V57.8576L61.437 58.2416L61.245 56.872L67.1202 56.616V55.1056H61.5522V53.8384ZM72.2149 53.8512H79.5109V55.1184H76.6181V55.464C76.6181 57.1536 77.6293 58.5872 79.8181 59.0992L79.0885 60.3024C77.5013 59.88 76.4133 58.984 75.8245 57.768C75.2485 58.984 74.1477 59.88 72.5605 60.3024L71.8309 59.0992C74.0325 58.5872 75.0309 57.1536 75.0309 55.464V55.1184H72.2149V53.8512Z"
-        fill="#95B0F8"
-      />
-      <path
-        d="M23.2002 45.6001L23.2002 68.0001"
-        stroke="#FF9100"
-        strokeWidth="1.6"
-        strokeLinecap="round"
-      />
-    </svg>
   )
 }
